test(Customer): Testes de CustomerService finalizados

+++ src/main/java/org/workshop/automanager/service/CustomerService.java
@@ -5,6 +5,8 @@ import org.springframework.stereotype.Service;
 import org.workshop.automanager.dto.request.CustomerRequestDTO;
 import org.workshop.automanager.dto.response.CustomerResponseDTO;
 import org.workshop.automanager.enums.RoleEnum;
+import org.workshop.automanager.exception.AlreadyExistsException;
+import org.workshop.automanager.exception.InvalidArgumentException;
 import org.workshop.automanager.exception.NotFoundException;
 import org.workshop.automanager.mapper.CustomerMapper;
 import org.workshop.automanager.model.CustomerEntity;
@@ -21,25 +23,33 @@ public class CustomerService {
     @Autowired
     private CustomerMapper customerMapper;
 
-    public void creteCustomer(CustomerRequestDTO requestDTO) throws Exception {
-        if (customerRepository.existsByCpf(requestDTO.getCpf()) || !requestDTO.getRole().equals(RoleEnum.CUSTOMER)) {
-            throw new Exception("Customer already registered!");
+    public void creteCustomer(CustomerRequestDTO requestDTO) {
+        if (!requestDTO.getRole().equals(RoleEnum.CUSTOMER)) {
+            throw new InvalidArgumentException("O Cargo " + requestDTO.getRole() + " recebido é inválido");
+        } else if (customerRepository.existsByCpf(requestDTO.getCpf())) {
+            throw new AlreadyExistsException("O clienta já está registrado.");
         }
         customerRepository.save(customerMapper.toCustomerEntity(requestDTO));
     }
 
-    public CustomerResponseDTO getCustomerByCpf(String cpf){
+    public CustomerResponseDTO getCustomerByCpf(String cpf) {
+        if (cpf == null) {
+            throw new InvalidArgumentException("O CPF " + cpf + " recebido é inválido");
+        }
         CustomerEntity entity = customerRepository.findByCpf(cpf);
         if (entity == null) {
-            throw new NotFoundException("Cliente não encontrado!");
+            throw new NotFoundException("O CPF " + cpf + " não foi encontrado!");
         }
         return customerMapper.toCustomerResponseDTO(entity);
     }
 
-    public CustomerResponseDTO getCustomerById(int id) throws Exception {
+    public CustomerResponseDTO getCustomerById(Integer id) {
+        if (id == null || id <= 0) {
+            throw new InvalidArgumentException("O ID " + id + " recebido é inválido");
+        }
         Optional<CustomerEntity> entity = customerRepository.findById(id);
         if (entity.isEmpty()) {
-            throw new Exception("Customer not found!");
+            throw new NotFoundException("Cliente não encontrado!");
         }
         return customerMapper.toCustomerResponseDTO(entity.get());
     }
@@ -48,10 +58,13 @@ public class CustomerService {
         return customerMapper.toCustomerResponseDTOList(customerRepository.findAll());
     }
 
-    public void updateCustomerByCpf(CustomerRequestDTO requestDTO) throws Exception {
+    public void updateCustomerByCpf(CustomerRequestDTO requestDTO) {
+        if (requestDTO.getCpf() == null || requestDTO.getCpf().isBlank()) {
+            throw new InvalidArgumentException("O CPF " + requestDTO.getCpf() + " recebido é inválido");
+        }
         CustomerEntity entity = customerRepository.findByCpf(requestDTO.getCpf());
         if (entity == null) {
-            throw new Exception("Customer not found!");
+            throw new NotFoundException("Cliente não encontrado!");
         }
         entity.setAddress(requestDTO.getAddress());
         entity.setName(requestDTO.getName());
@@ -60,10 +73,13 @@ public class CustomerService {
         customerRepository.save(entity);
     }
 
-    public void deleteByCpf(String cpf) throws Exception {
+    public void deleteByCpf(String cpf) {
+        if (cpf == null || cpf.isBlank()) {
+            throw new InvalidArgumentException("O CPF " + cpf + " recebido é inválido");
+        }
         CustomerEntity entity = customerRepository.findByCpf(cpf);
         if (entity == null) {
-            throw new Exception("Customer not found!");
+            throw new NotFoundException("Cliente não encontrado!");
         }
         customerRepository.delete(entity);
     }
